FourierTools.jl

Here you can find the docstrings of all functions.

FFT based resampling

This allows to sinc interpolate a signal. The resulting size can be larger or smaller.

FourierTools.resample — Function

resample(arr, new_size [, normalize])

Calculates the sinc interpolation of an arr on a new array size new_size. It is a re-evaluation of the Fourier series at new grid points. new_size can be arbitrary. Each dimension is then independently either up or downsampled. This method is based on FFTs and therefore implicitly assumes periodic boundaries and a finite frequency support. normalize=true by default multiplies by an appropriate factor so that the average intensity stays the same. If size(new_size)[i] > size(arr)[i], we apply zero padding in Fourier space. If size(new_size)[i] < size(arr)[i], we cut out a centered part of the Fourier spectrum. We apply some tricks at the boundary to increase accuracy of highest frequencies. ```

FFT helpers

FourierTools.ft — Function

ft(A [, dims])

Result is semantically equivalent to fftshift(fft(A, dims), dims) However, the shift is done with ShiftedArrays and therefore doesn't allocate memory.

FourierTools.ift — Function

ift(A [, dims])

Result is semantically equivalent to ifft(ifftshift(A), dims), dims)

FourierTools.rft — Function

rft(A [, dims])

Calculates a rfft(A, dims) and then shift the frequencies to the center. dims[1] is not shifted, because there is no negative and positive frequency. The shift is done with ShiftedArrays and therefore doesn't allocate memory.

FourierTools.irft — Function

irft(A, d, [, dims])

Calculates a irfft(A, d, dims) and then shift the frequencies back to the corner. dims[1] is not shifted, because there is no negative and positive frequency. The shift is done with ShiftedArrays and therefore doesn't allocate memory.

FourierTools.fftshift_view — Function

fftshift_view(A [, dims])

Result is semantically equivalent to fftshift(A, dims) but returns a view instead.

FourierTools.ifftshift_view — Function

ifftshift_view(A [, dims])

Result is semantically equivalent to fftshift(A, dims) but returns a view instead.