FourierTools.jl

Here you can find the docstrings of all functions.

FFT based resampling

This allows to sinc interpolate a signal. The resulting size can be larger or smaller.

FourierTools.resample — Function

resample(arr, new_size [, normalize])

Calculates the sinc interpolation of an arr on a new array size new_size. It is a re-evaluation of the Fourier series at new grid points. new_size can be arbitrary. Each dimension is then independently either up or downsampled. This method is based on FFTs and therefore implicitly assumes periodic boundaries and a finite frequency support. normalize=true by default multiplies by an appropriate factor so that the average intensity stays the same. If size(new_size)[i] > size(arr)[i], we apply zero padding in Fourier space. If size(new_size)[i] < size(arr)[i], we cut out a centered part of the Fourier spectrum. We apply some tricks at the boundary to increase accuracy of highest frequencies. ```

FFT helpers

FourierTools.ft — Function

ft(A [, dims])

Digital Fourier-transformation centered in both spaces. The result is semantically equivalent to fftshift(fft(ifftshift(A, dims), dims), dims) This is a digital Fourier transformation with both coordinate systems in real and Fourier-space being centered at position CtrFT == size÷2+1

The following identities are true: ft(ones(sz)) ≈ prod(sz) .* delta(sz) ft(delta(sz)) ≈ ones(sz)

#see also: ift, rft, irft, ffts, iffts However, the shift is done with ShiftedArrays and therefore doesn't allocate memory.

FourierTools.ift — Function

ift(A [, dims])

Digital inverse Fourier-transformation centered in both spaces.
The result is semantically equivalent to `fftshift(ifft(ifftshift(A, dims), dims), dims)`
This is a digital Fourier transformation with both coordinate systems in real and Fourier-space being
centered at position CtrFT == size÷2+1

The following identities are true:
ift(ones(sz)) ≈ delta(sz)
ift(delta(sz)) ≈ ones(sz) ./ prod(sz)

#see also: ft, rft, irft, ffts, iffts

FourierTools.rft — Function

rft(A [, dims])

Digital real-valued Fourier-transformation centered in both spaces.
The result is semantically equivalent to `fftshift(rfft(ifftshift(A, dims), dims), dims)`
This is a digital Fourier transformation with the coordinate systems in real space centered at CtrFT == size÷2+1
and in (half) Fourier-space being centered at CtrRFT == setindex(size÷2 +1,1,1).

The following identities are true:
rft(ones(sz)) ≈ prod(sz) .* delta(rft_size(sz), offset=CtrRFT)
rft(delta(sz)) ≈ ones(rft_size(sz))

#see also: ft, ift, irft, ffts, iffts

FourierTools.irft — Function

irft(A, d, [, dims])

Digital real-valued inverse Fourier-transformation centered in both spaces.
The result is semantically equivalent to `fftshift(irfft(ifftshift(A, dims), dims), dims)`
This is a digital Fourier transformation with the coordinate systems in real space centered at CtrFT == size÷2+1
and in (half) Fourier-space being centered at CtrRFT == setindex(size÷2 +1,1,1).
Note that the size `d` of the first transform direction [1] is a required argument.

The following identities are true:
irft(ones(rft_size(sz)),sz[1]) ≈ delta(sz)
irft(delta(rft_size(sz),offset=CtrRFT),sz[1]) ≈ ones(sz) ./ prod(sz)

#see also: ft, ift, rft, ffts, iffts

FourierTools.fftshift_view — Function

fftshift_view(A [, dims])

Result is semantically equivalent to fftshift(A, dims) but returns a view instead.

FourierTools.ifftshift_view — Function

ifftshift_view(A [, dims])

Result is semantically equivalent to fftshift(A, dims) but returns a view instead.